PHPStan level 5

+++ models/query/BattleQuery.php
@@ -223,15 +223,15 @@ final class BattleQuery extends ActiveQuery
                 break;
 
             case 'today':
-                $t = mktime(0, 0, 0, date('n', $now), date('j', $now), date('Y', $now));
+                $t = mktime(0, 0, 0, (int)date('n', $now), (int)date('j', $now), (int)date('Y', $now));
                 $this->andWhere(['>=', '{{battle}}.[[at]]', gmdate('Y-m-d\TH:i:sP', $t)]);
                 break;
 
             case 'yesterday':
                 // 昨日の 00:00:00
-                $t1 = mktime(0, 0, 0, date('n', $now), date('j', $now) - 1, date('Y', $now));
+                $t1 = mktime(0, 0, 0, (int)date('n', $now), (int)date('j', $now) - 1, (int)date('Y', $now));
                 // 今日の 00:00:00
-                $t2 = mktime(0, 0, 0, date('n', $now), date('j', $now), date('Y', $now));
+                $t2 = mktime(0, 0, 0, (int)date('n', $now), (int)date('j', $now), (int)date('Y', $now));
                 $this->andWhere(['>=', '{{battle}}.[[at]]', gmdate('Y-m-d\TH:i:sP', $t1)]);
                 $this->andWhere(['<', '{{battle}}.[[at]]', gmdate('Y-m-d\TH:i:sP', $t2)]);
                 break;
